[#56]: Prepare backend for development
Problem:
Backend app was very empty, with default definitions from template. We
need to prepare it for development process.

Solution:
This commit is about defining initial architecture, so as result we
have some monads to work with.
For now we have application monad where we can read config at startup
and store it in env along with DB pool connection.
Also we have simple API to return current config, it will be deleted and
now exists only to show that we can server API.
We have some initial error types and some core definitions to use
monitoring and jwt (jwt tests presented).