[Chore]: Change code formatter
Problem:
For now it's not hard to change code formatter because of small code
base, in future it will be harder.
Currently we use `stylish-haskell`, which mostly formats only imports.
Sometimes it's not enough and you can miss some style requirements.

Solution:
Change formatter to `brittany`, provide config for it and apply
formatting.