[#69]: Provide migrations mechanism
Problem:
We need to track changes at database, to achieve this - we need
migrations mechanism for our backend, so we can be sure, that all
required DB changes will be applied at app startup and only once. But
in case when smth goes wrong, e.g. migration script was changed, we want
to get an error.

Solution:
* Add helper functions to create transactions and execute it, provide
  function to initialize db with schema to store migrations history and
  execute custom migrations form dir (you can set it up in config).
* Provide some helper functions to work with errors - this is needed
  to properly translate database error to app internal one, and then
  print it pretty.
* Provide command-line interface to create migrations, list it and prune
  DB.