[#76]: Fix logs printing in multiple threads
Problem:
Function to print text with newline which is in use in library performs
it as two separate actions - first is to print message and second - to
print newline symbol.
This way there is issues with computations in multiple threads, so
newline symbol sometimes printed in wrong place.

Solution:
Redefine logger action to use version where newline symbol appended
before printing.