[#76]: Add logging middleware
Problem:
This is first step to log http requests. Now we want to actually log
requests to understand what is going on at http layer.

Solution:
Add simple logger without customizations, it will log on `Debug` and
`Info` severity only, but with different verbosity.
In future we can customize it or even use same logger as for whole app,
but for now this effort is not required.