[#82]: Add deployment pipeline
Problem:
So, now we have all parts to build and ship our backend app and we need
to make last step - upload some files to server (and keep up to date)
and relaunch app with new version.

Solution:
Add github action for deployment which uses github environment and ssh
(and rsync) to upload docker staff to remote and run `docker compose`
there.