[Chore]: Use benefits of new version of servant
Problem:
Before we had to map record definition of API to servant terms with
some helper functions, but now we can do it first-class.

Solution:
Drop `toServant` and related and use `NamedRoutes` instead; simplify
using client in tests - now we have helper function which refers to
whole api and can just `//` to point to required endpoint.